+++ app/Http/Resources/InvoiceItem/InvoiceItemResource.php
@@ -22,21 +22,21 @@ class InvoiceItemResource extends JsonResource
             'creator_id' => $this->creator_id,
             'item_id' => $this->item_id,
             'user_id' => $this->user_id,
-            'returned_qty' => (int)$this->returned_qty,
-            'qty' => (int)$this->qty,
+            'returned_qty' => (float)$this->returned_qty,
+            'qty' => (float)$this->qty,
             'tax' => (float)($this->tax),
             'price' => (float)($this->price),
             'total' => (float)($this->total),
             'subtotal' => (float)($this->subtotal),
             'discount' => (float)($this->discount),
-            'unit_price' => $this->qty == 0  ? 0: (float)($this->subtotal / $this->qty) ,
+            'unit_price' => $this->qty == 0  ? 0 : (float)($this->subtotal / $this->qty),
             'invoice_type' => $this->invoice_type,
             'belong_to_kit' => $this->belong_to_kit,
             'parent_kit_id' => $this->parent_kit_id,
             'is_kit' => $this->is_kit,
             'show_price_in_print_mode' => (float)$this->show_price_in_print_mode,
             'is_draft' => (bool)$this->is_draft,
-            'available_qty' => (int)$this->available_qty,
+            'available_qty' => (float)$this->available_qty,
             'cost' => (float)($this->cost),
             'total_stock_cost_amount' => ((float)$this->total_stock_cost_amount),
             'profit' => (float)$this->profit,
@@ -52,6 +52,6 @@ class InvoiceItemResource extends JsonResource
 
 
         ];
-//        return parent::toArray($request);
+        //        return parent::toArray($request);
     }
 }
